Fix for XML parsing error when viewing UIKit namespace summary
The problem was the use of a "<" in the phrase "only necessary in  iOS <6" in `UITableView.xml`, line 195, which ultimately becomes line 3469 in `xml.summary.MonoTouch.UIKit`.

But what I don't understand is that this character occurs inside a <![CDATA[]]> section, which shouldn't be parsed by `XmlTextReader.ReadName`.

Also confusing is that when I generate MonoTouch-lib.zip on my machine from the *same source*, I get a file that can be opened in `macdoc`. Is it possible that the `mdoc` that generates the MonoTouch-lib.zip for distribution is not current?

The May 9 mono checkin 245dee37992a827b99b9cb9df36cf78ce65db454 seems highly relevant, since the remarks element in these summary files *used to be* automatically made to be CDATA.