Hotfix to avoid monodoc bug that does not properly ignore CDATA'd XML
Although this code section is wrapped in CDATA, the text "List<PageController>" leads to trouble in the MonoTouch.UIKit namespace page, at the point where it is building the list of member types.

This is probably the result of a regex or text-extraction that is not properly excluding CDATA elements.

There are a *lot* of <code>...</code> elements whose contents are not wrapped in CDATA (yet). Until we address that, our XML parsing is always going to have a certain funk to it.