Docs for GetBestMove
Added paragraph that explains why a GKMinMaxStrategist playing TicTacToe will not return the "optimal" move (a corner), but instead will randomly return *any* cell: even if the first X play is to the center or an edge, it can *still* force a tie. So *every* cell on a blank TicTacToe board has the same expected value: "tie".