AVCaptureSession code changes; fixes https://github.com/xamarin/documentation/issues/1570
Fix for https://github.com/xamarin/documentation/issues/1570

Two problems with the existing code: one was missing code for `AVCaptureDeviceInput` and the other was an explicit call that the info.plist _must_ contain the NSMicrophoneUsageDescription or the app crashes.

Also swept through other code samples in type-level <remarks/> for formatting and confirming them against live code.